Carousel: fix portrait black screen when gallery image has no data-large-file
Gallery images that were never enriched with Jetpack's data-large-file
attribute reach selectBestImageUrl() with largeFile === '' (a missing
attribute reads as an empty string). In a narrow, portrait-mobile viewport
the 'large file is big enough' guard fired and returned that empty string
as the slide's image source, leaving the carousel with a blank (black)
slide. Landscape and desktop viewports are wide enough that the guard is
never taken, which is why the bug only surfaced on mobile in portrait.

Guard against a falsy largeFile (not just undefined) so it falls back to
the original file in every orientation. The URL-selection helpers are
extracted into select-best-image-url.js so the logic can be unit tested.
